+++ src/test/java/org/jabref/logic/integrity/BooktitleCheckerTest.java
@@ -3,8 +3,6 @@ package org.jabref.logic.integrity;
 import java.util.Optional;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
@@ -26,20 +24,5 @@ class BooktitleCheckerTest {
     @Test
     void booktitleIsBlank() {
         assertEquals(Optional.empty(), checker.checkValue(" "));
-    }
-
-    @Test
-    void booktitleShouldAcceptURLWithoutProtocol() {
-        assertEquals(Optional.empty(), checker.checkValue("Applying {T}rip@dvice {R}ecommendation {T}echnology to www.visiteurope.com"));
-    }
-
-    @ParameterizedTest(name = "{index}. Booktitle: \"{0}\" should be invalid")
-    @CsvSource({
-            "Proceedings of the https://example.com/conference",
-            "Find more at http://mywebsite.org/article",
-            "Visit ftp://files.example.com/download",
-    })
-    void booktitleShouldRaiseWarning(String booktitle) {
-        assertNotEquals(Optional.empty(), checker.checkValue(booktitle));
     }
 }