+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Note that this project **does not** adhere to [Semantic Versioning](https://semv
 ### Added
 
 - We added a feature for copying entries to libraries, available via the context menu, with an option to include cross-references. [#12374](https://github.com/JabRef/jabref/pull/12374)
-- We added an error message that appears when a user tries to enter a URL as a book title. [#12354](https://github.com/JabRef/jabref/issues/12354)
+- We added an integrity check if a URL appears in a title. [#12354](https://github.com/JabRef/jabref/issues/12354)
 
 ### Changed